LPS-135490 Moves creation menu logic for widget templates to corresponding class
