+++ modules/apps/template/template-web/src/main/java/com/liferay/template/web/internal/display/context/BaseTemplateManagementToolbarDisplayContext.java
@@ -14,10 +14,8 @@
 
 package com.liferay.template.web.internal.display.context;
 
-import com.liferay.dynamic.data.mapping.constants.DDMTemplateConstants;
 import com.liferay.dynamic.data.mapping.model.DDMTemplate;
 import com.liferay.frontend.taglib.clay.servlet.taglib.display.context.SearchContainerManagementToolbarDisplayContext;
-import com.liferay.frontend.taglib.clay.servlet.taglib.util.CreationMenu;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemListBuilder;
 import com.liferay.petra.portlet.url.builder.PortletURLBuilder;
@@ -34,7 +32,6 @@ import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.template.web.internal.security.permissions.resource.DDMTemplatePermission;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import javax.portlet.PortletURL;
@@ -104,57 +101,6 @@ public abstract class BaseTemplateManagementToolbarDisplayContext
 	}
 
 	@Override
-	public CreationMenu getCreationMenu() {
-		if (!templateDisplayContext.isAddDDMTemplateEnabled()) {
-			return null;
-		}
-
-		List<Long> addAllowedClassNameIds = _getAddAllowedClassNameIds();
-
-		if (addAllowedClassNameIds.isEmpty()) {
-			return null;
-		}
-
-		CreationMenu creationMenu = new CreationMenu();
-
-		PortletURL addDDMTemplateURL = PortletURLBuilder.createRenderURL(
-			liferayPortletResponse
-		).setMVCPath(
-			"/edit_ddm_template.jsp"
-		).setRedirect(
-			themeDisplay.getURLCurrent()
-		).setTabs1(
-			templateDisplayContext.getTabs1()
-		).setParameter(
-			"groupId", themeDisplay.getScopeGroupId()
-		).setParameter(
-			"type", DDMTemplateConstants.TEMPLATE_TYPE_DISPLAY
-		).buildPortletURL();
-
-		for (long addAllowedClassNameId : addAllowedClassNameIds) {
-			addDDMTemplateURL.setParameter(
-				"classNameId", String.valueOf(addAllowedClassNameId));
-			addDDMTemplateURL.setParameter("classPK", "0");
-			addDDMTemplateURL.setParameter(
-				"resourceClassNameId",
-				String.valueOf(
-					templateDisplayContext.getResourceClassNameId()));
-
-			creationMenu.addPrimaryDropdownItem(
-				dropdownItem -> {
-					dropdownItem.setHref(addDDMTemplateURL);
-					dropdownItem.setLabel(
-						LanguageUtil.get(
-							httpServletRequest,
-							templateDisplayContext.getTemplateTypeLabel(
-								addAllowedClassNameId)));
-				});
-		}
-
-		return creationMenu;
-	}
-
-	@Override
 	public String getDefaultEventHandler() {
 		return "TEMPLATE_MANAGEMENT_TOOLBAR_DEFAULT_EVENT_HANDLER";
 	}
@@ -199,18 +145,6 @@ public abstract class BaseTemplateManagementToolbarDisplayContext
 
 	protected final TemplateDisplayContext templateDisplayContext;
 	protected final ThemeDisplay themeDisplay;
-
-	private List<Long> _getAddAllowedClassNameIds() {
-		List<Long> addAllowedClassNameIds = new ArrayList<>();
-
-		for (long classNameId : templateDisplayContext.getClassNameIds()) {
-			if (containsAddPortletDisplayTemplatePermission(classNameId)) {
-				addAllowedClassNameIds.add(classNameId);
-			}
-		}
-
-		return addAllowedClassNameIds;
-	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		BaseTemplateManagementToolbarDisplayContext.class);
