+++ modules/apps/template/template-web/src/main/java/com/liferay/template/web/internal/display/context/WidgetTemplatesManagementToolbarDisplayContext.java
@@ -14,8 +14,17 @@
 
 package com.liferay.template.web.internal.display.context;
 
+import com.liferay.dynamic.data.mapping.constants.DDMTemplateConstants;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.CreationMenu;
+import com.liferay.petra.portlet.url.builder.PortletURLBuilder;
+import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import javax.portlet.PortletURL;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -34,6 +43,69 @@ public class WidgetTemplatesManagementToolbarDisplayContext
 		super(
 			httpServletRequest, liferayPortletRequest, liferayPortletResponse,
 			templateDisplayContext);
+	}
+
+	@Override
+	public CreationMenu getCreationMenu() {
+		if (!templateDisplayContext.isAddDDMTemplateEnabled()) {
+			return null;
+		}
+
+		List<Long> addAllowedClassNameIds = _getAddAllowedClassNameIds();
+
+		if (addAllowedClassNameIds.isEmpty()) {
+			return null;
+		}
+
+		CreationMenu creationMenu = new CreationMenu();
+
+		PortletURL addDDMTemplateURL = PortletURLBuilder.createRenderURL(
+			liferayPortletResponse
+		).setMVCPath(
+			"/edit_ddm_template.jsp"
+		).setRedirect(
+			themeDisplay.getURLCurrent()
+		).setTabs1(
+			templateDisplayContext.getTabs1()
+		).setParameter(
+			"groupId", themeDisplay.getScopeGroupId()
+		).setParameter(
+			"type", DDMTemplateConstants.TEMPLATE_TYPE_DISPLAY
+		).buildPortletURL();
+
+		for (long addAllowedClassNameId : addAllowedClassNameIds) {
+			addDDMTemplateURL.setParameter(
+				"classNameId", String.valueOf(addAllowedClassNameId));
+			addDDMTemplateURL.setParameter("classPK", "0");
+			addDDMTemplateURL.setParameter(
+				"resourceClassNameId",
+				String.valueOf(
+					templateDisplayContext.getResourceClassNameId()));
+
+			creationMenu.addPrimaryDropdownItem(
+				dropdownItem -> {
+					dropdownItem.setHref(addDDMTemplateURL);
+					dropdownItem.setLabel(
+						LanguageUtil.get(
+							httpServletRequest,
+							templateDisplayContext.getTemplateTypeLabel(
+								addAllowedClassNameId)));
+				});
+		}
+
+		return creationMenu;
+	}
+
+	private List<Long> _getAddAllowedClassNameIds() {
+		List<Long> addAllowedClassNameIds = new ArrayList<>();
+
+		for (long classNameId : templateDisplayContext.getClassNameIds()) {
+			if (containsAddPortletDisplayTemplatePermission(classNameId)) {
+				addAllowedClassNameIds.add(classNameId);
+			}
+		}
+
+		return addAllowedClassNameIds;
 	}
 
 }
